fix: prevent authenticated access via browser back button after logout
